W432: 1 function matched, first-party nd -598, and a closed defect hypothesis
First-party MATCH 6,887 -> 6,888 of 8,235 (83.6%).
Total MATCH 8,053 -> 8,054. NONMATCHING 1,347 -> 1,346.
Zero MATCHes lost, zero blocking statuses, zero new over-window functions,
zero regressions.

Newly matching
  00227800  Main/Battle/Panel/bp_tex.c  nd 157 -> 0

Largest reductions short of a match
  003dcfb0  Fcl/fclCombineList.c   371 -> 292
  00413010  Fcl/fclShopBGManager.c 153 ->  69
  00351a10  Graphics/Model/mdlEffect.c  223 -> 161
  003e6e80  Fcl/fclShop.c         3327 -> 3274
  003ef2b0  Fcl/fclShop.c          957 -> 922
  003005e0  Battle/Data/datCalc.c   93 ->  31
  001ce960  Kosaka/Field/k_unit.c 2894 -> 2866
  0042bc10  Yajima/y_misc.c         61 ->  45

This wave tested a hypothesis and it failed. A corpus census reported 3,918
inline scalar casts inside call argument lists across first-party NONMATCHING
functions, on the theory that a redundant cast forces MWCC to materialise an
argument earlier than retail does. Sixteen lanes tested it. The result is a
clean negative: essentially all of those sites are pointer conversions, genuine
narrowing or signedness changes, or int/float conversions that the code needs.
Roughly 60 removals were attempted project-wide; none improved a single
function, several failed to compile, and all were reverted. The same census
claimed 215 un-suffixed float literals; almost all were text inside block
comments, and the real source was already correctly suffixed. Four genuine ones
in br_data.c were corrected.

Recording that plainly so no future wave repeats it. Last wave's datCalendar
result came from the position-first sprite alias changing parameter ORDER, not
from the cast removals that accompanied it.

What did work, once lanes fell back to the ranked method, was the same set of
techniques as the previous three waves: local declaration order to control
callee-saved register assignment (worth 53 and 35 points on two fclShop
functions and 79 on fclCombineList), guard predicate correction where retail
emits blez rather than beqz because the original compared with an ordering
operator, stack aggregate extent correction, and ascending-order switch
dispatch with the body moved out of line.

One durable rule came out of it, and it is worth more than the wave's nd.
When a function's object_size EQUALS its retail window, our build and retail
emit the same instruction count, so any transformation that changes that count
is wrong by construction. Only size-preserving moves can be correct there:
guard polarity, declaration order, splitting or merging locals, commutative
operand order. Lanes burned several probes converting exact-size if-chains into
switches, which adds a comparison and immediately overruns the window. The size
is checkable before compiling.

Verified: tools/build.py reproduces SLUS_216.21 sha1
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb; 92 tests pass.

+++ src/Battle/btlAction.c
@@ -2627,7 +2627,7 @@ check_distance_threshold:
 
 compute_vector_offset:
     {
-        FUN_0027f940(personaUnit, victimUnit, actionUnit, (s32)(u32)personaActionType, &sp110, NULL, 2);
+        FUN_0027f940(personaUnit, victimUnit, actionUnit, personaActionType, &sp110, NULL, 2);
 
         spE0.x = sp110.x - spE0.x;
         spE0.y = 0.0f;
@@ -4409,7 +4409,7 @@ void btlActionUpdateStateSkill(BtlAction *action) {
                 temp_v0_18->parentUID = temp_v0_17->uid;
                 temp_v0_18->actionUID = temp_s0;
                 btlPacketRegister(temp_v0_18, 1U);
-                temp_v0_19 = btlVoice002e2be0(action, 0xC, (s32) sp450->charId, 0, 0);
+                temp_v0_19 = btlVoice002e2be0(action, 0xC, sp450->charId, 0, 0);
                 temp_v0_19->unk_00 = 5;
                 temp_v0_19->parentUID = temp_v0_18->uid;
                 temp_v0_19->actionUID = temp_s0;
@@ -5179,7 +5179,7 @@ loop_302:
             temp_v0_116 = (u8*)temp_s2_3 + sp360 * 0x1C;
             sp110 = (u64) temp_v0_116;
             sp100 = (u64) ((u8*)temp_v0_116 + 0xF8);
-            temp_v0_117 = btlUnitCreateAnimPacket(var_s5_2->unit, (u16) ACTION_U8(temp_s2_3, 0xf8 + sp360 * 0x1c), 0U, 1.0f, 0U);
+            temp_v0_117 = btlUnitCreateAnimPacket(var_s5_2->unit, ACTION_U8(temp_s2_3, 0xf8 + sp360 * 0x1c), 0U, 1.0f, 0U);
             var_s3 = temp_v0_117;
             temp_v0_117->unk_00 = 0xB;
             temp_v0_117->parentUID = var_s1_3->uid;
@@ -5448,7 +5448,7 @@ block_325:
             }
         }
     } else if ((sp380 > 0) && (temp_v1_14 = ACTION_U8(action, 0xca), ((temp_v1_14 & 1) != 0)) && !(temp_v1_14 & 2)) {
-        temp_v0_144 = func_002db740(action, 0x17, (s32) sp320, 0);
+        temp_v0_144 = func_002db740(action, 0x17, sp320, 0);
         temp_v0_144->unk_00 = 5;
         temp_v0_144->parentUID = temp_s2_4;
         temp_v0_144->preUpdateDelay = (s16) sp3C0;
@@ -5669,7 +5669,7 @@ block_394:
             temp_v0_164->preUpdateDelay = var_v0_12;
             temp_v0_164->actionUID = temp_s0;
             btlPacketRegister(temp_v0_164, 1U);
-            temp_v0_165 = func_002bb2f0((*(u32*)((u8*)gBtl + 0xCEC)), sp450, (u32) sp3B0, 0U, 0);
+            temp_v0_165 = func_002bb2f0((*(u32*)((u8*)gBtl + 0xCEC)), sp450, sp3B0, 0U, 0);
             temp_v0_165->unk_00 = 4;
             temp_v0_165->parentUID = sp420;
             temp_v0_165->actionUID = temp_s0;
